bigdecimal-double

+++ src/main/java/com/example/backend_nutripoint/DTO/ProductResponseDTO.java
@@ -1,8 +1,12 @@
 package com.example.backend_nutripoint.DTO;
 
+import java.math.BigDecimal;
 import java.util.List;
 
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
 
 @Getter
 @Setter
@@ -14,7 +18,7 @@ public class ProductResponseDTO {
     private String descripcion;
     private Integer stock;
     private String marca;
-    private Double preciounit;
+    private BigDecimal preciounit;
     private String modEmpleo;
     private String advert;
     private List<String> imagenesUrls;
